requirements.txt
reflex==0.6.8
python-decouple
openai
psycopg2-binary